Implement Ezra's suggestion
src/mongocxx/test/uri.cpp is reserved for URI tests that do not require
a live server.

For consistency with other unified spec tests, move these tests into a
new src/mongocxx/test/spec/uri.cpp file.
